Move middleware/tls to middleware/pkg/tls
This was put in the wrong place originally.
